requirements.txt
requests
gmusicapi
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# awscli 1.16.159 requirements:
rsa==3.4.2
# botocore 1.12.149 requirements:
urllib3==1.24.1
